scripts/python/requirements.txt
psutil>=5.9.0
rich>=13.0.0
pyyaml>=6.0
click>=8.0.0
requests>=2.28.0
gitpython>=3.1.0
